Add LotTransaction audit trail fields and fix unpick workflow
Backend Changes:
- Add PackSize retrieval from cust_PartialPicked for Phase 1 audit trail
- Fix SMALLINT type mapping: use i16 for ReceiptDocLineNo and IssueDocLineNo
- Fix DateTime binding: unwrap Option<DateTime> before Tiberius INSERT
- Populate complete audit trail in Cust_PartialLotPicked (ReceiptDocNo, Vendorkey, VendorlotNo)
- Add descriptive audit field values (ProcessReason: "Picking Customization")
- Fix unpick Phase 1: preserve audit trail by only resetting PickedPartialQty=0

Type Mapping Fixes:
- SQL Server SMALLINT → Rust i16 (was incorrectly using i32)
- SQL Server FLOAT(53) → Rust f64 (document common f32 mistake)
- Document Tiberius type conversion patterns for datetime columns

Documentation:
- Add Pitfall 8 to DB-Flow.md: Missing audit trail fields
- Add SQL Server to Rust type mapping reference to CLAUDE.md
- Document common type conversion mistakes and solutions

+++ backend/src/models/transaction.rs
@@ -38,9 +38,9 @@ pub struct LotTransaction {
     #[serde(rename = "ReceiptDocNo")]
     pub receipt_doc_no: Option<String>,
 
-    /// Receipt line
+    /// Receipt line (SQL Server SMALLINT)
     #[serde(rename = "ReceiptDocLineNo")]
-    pub receipt_doc_line_no: Option<i32>,
+    pub receipt_doc_line_no: Option<i16>,
 
     /// Qty received (0 for picking)
     #[serde(rename = "QtyReceived")]
@@ -58,9 +58,9 @@ pub struct LotTransaction {
     #[serde(rename = "IssueDocNo")]
     pub issue_doc_no: Option<String>,
 
-    /// Issue line (LineId from cust_PartialPicked)
+    /// Issue line (LineId from cust_PartialPicked, SQL Server SMALLINT)
     #[serde(rename = "IssueDocLineNo")]
-    pub issue_doc_line_no: Option<i32>,
+    pub issue_doc_line_no: Option<i16>,
 
     /// Picking timestamp
     #[serde(rename = "IssueDate")]
